Decimal support on GeoLocation options

+++ lib/PuppeteerSharp/GeolocationOption.cs
@@ -13,17 +13,17 @@ namespace PuppeteerSharp
         /// Latitude between -90 and 90.
         /// </summary>
         /// <value>The latitude.</value>
-        public int Latitude { get; set; }
+        public decimal Latitude { get; set; }
         /// <summary>
         /// Longitude between -180 and 180.
         /// </summary>
         /// <value>The longitude.</value>
-        public int Longitude { get; set; }
+        public decimal Longitude { get; set; }
         /// <summary>
         /// Optional non-negative accuracy value.
         /// </summary>
         /// <value>The accuracy.</value>
-        public int Accuracy { get; set; }
+        public decimal Accuracy { get; set; }
 
         /// <summary>
         /// Determines whether the specified <see cref="PuppeteerSharp.GeolocationOption"/> is equal to the current <see cref="T:PuppeteerSharp.GeolocationOption"/>.
